# server/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "DCRM API",
    "description": "API REST du CRM interne DCB Technologies. Format de réponse unifié : succès `{ success: true, data, meta? }`, erreur `{ success: false, error: { code, message } }`. Authentification par JWT Bearer (15 min, refresh 7 j via cookie httpOnly) ou par clé API longue durée (header X-API-Key). L'extension `x-permission` sur une opération indique la clé de permission RBAC requise. Tous les montants monétaires sont exprimés en euros hors taxes (HT) ; le TTC n'est jamais stocké et se dérive via Product.vatRate (défaut 20 %).",
    "version": "2.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3001/api",
      "description": "Développement"
    },
    {
      "url": "https://{domaine}/api",
      "description": "Production",
      "variables": {
        "domaine": {
          "default": "crm.example.fr"
        }
      }
    }
  ],
  "security": [
    {
      "bearerAuth": []
    },
    {
      "apiKeyAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Access token obtenu via POST /auth/login (durée 15 min)"
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Clé API longue durée générée via POST /apikeys (préfixe dcrm_). Permissions effectives = intersection des droits de la clé (scopes) et des droits actuels de son propriétaire — jamais de bypass '*', même pour un ADMIN ; les routes gardées par rôle sont inaccessibles par clé API."
      }
    },
...
  },
  "paths": {
...
    "/products": {
      "get": {
        "tags": [
          "Produits"
        ],
        "summary": "Lister les produits (paginé, filtrable)",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ],
        "x-permission": "products:read",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "category",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "isActive",
            "in": "query",
            "description": "'true' = actifs seuls, 'false' = inactifs seuls, absent = tous",
            "schema": {
              "type": "string",
              "enum": ["true", "false"]
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 50,
              "maximum": 200
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Liste paginée des produits"
          },
          "401": {
            "description": "Non authentifié"
          },
          "403": {
            "description": "Permission products:read manquante"
          }
        }
      },
      "post": {
        "tags": [
          "Produits"
        ],
        "summary": "Créer un produit",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ],
        "x-permission": "products:create",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "category",
                  "price"
                ],
                "properties": {
                  "reference": {
                    "type": "string"
                  },
                  "name": {
                    "type": "string",
                    "minLength": 1
                  },
                  "description": {
                    "type": "string"
                  },
                  "category": {
                    "type": "string"
                  },
                  "type": {
                    "type": "string"
                  },
                  "price": {
                    "type": "number"
                  },
                  "vatRate": {
                    "type": "number"
                  },
                  "unit": {
                    "type": "string"
                  },
                  "stock": {
                    "type": "integer"
                  },
                  "supplier": {
                    "type": "string"
                  },
                  "imageUrl": {
                    "type": "string"
                  },
                  "isActive": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Produit créé"
          },
          "400": {
            "description": "VALIDATION_ERROR"
          },
          "401": {
            "description": "Non authentifié"
          },
          "403": {
            "description": "Permission products:create manquante"
          }
        }
      }
    },
    "/products/{id}": {
      "get": {
        "tags": [
          "Produits"
        ],
        "summary": "Détail d'un produit",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ],
        "x-permission": "products:read",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Produit"
          },
          "401": {
            "description": "Non authentifié"
          },
          "403": {
            "description": "Permission products:read manquante"
          },
          "404": {
            "description": "NOT_FOUND"
          }
        }
      },
      "put": {
        "tags": [
          "Produits"
        ],
        "summary": "Mettre à jour un produit",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ],
        "x-permission": "products:update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "reference": {
                    "type": "string"
                  },
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "category": {
                    "type": "string"
                  },
                  "type": {
                    "type": "string"
                  },
                  "price": {
                    "type": "number"
                  },
                  "vatRate": {
                    "type": "number"
                  },
                  "unit": {
                    "type": "string"
                  },
                  "stock": {
                    "type": "integer"
                  },
                  "supplier": {
                    "type": "string"
                  },
                  "imageUrl": {
                    "type": "string"
                  },
                  "isActive": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Produit mis à jour"
          },
          "400": {
            "description": "VALIDATION_ERROR"
          },
          "401": {
            "description": "Non authentifié"
          },
          "403": {
            "description": "Permission products:update manquante"
          },
          "404": {
            "description": "NOT_FOUND"
          }
        }
      },
      "delete": {
        "tags": [
          "Produits"
        ],
        "summary": "Désactiver un produit (soft delete, isActive=false — pas de suppression réelle)",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ],
        "x-permission": "products:delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Produit désactivé"
          },
          "401": {
            "description": "Non authentifié"
          },
          "403": {
            "description": "Permission products:delete manquante"
          },
          "404": {
            "description": "NOT_FOUND"
          }
        }
      }
    },
...
  }
}